Separate TextUnitInput/convert.js
