+++ src/web/forms/fields/TextUnitInput/__tests__/index-test.js
@@ -18,72 +18,12 @@
 
 jest.unmock('react-redux');
 jest.unmock('../index');
+jest.unmock('../convert');
 
 import React from 'react';
 import { shallow } from 'enzyme';
-import math from 'lib/mathjs';
 
-import { convert, TextUnitInputComponent } from '../index';
-
-describe('TextUnitInput.convert', () => {
-  let MockMathUnit;
-  let mockMathUnitToNumber;
-
-  beforeEach(() => {
-    mockMathUnitToNumber = jest.fn().mockReturnValue(1234);
-    MockMathUnit = class {
-      toNumber = mockMathUnitToNumber;
-    };
-    math.unit = jest.fn().mockReturnValue(new MockMathUnit());
-  });
-
-  it('converts unit', () => {
-    const value = {
-      value: '180',
-      unit: 'cm',
-    };
-    const targetUnit = 'in';
-
-    convert(value, targetUnit);
-
-    expect(math.unit).toBeCalledWith('180', 'cm');
-    expect(mockMathUnitToNumber).toBeCalledWith('in');
-  });
-
-  it('returns as invalid input given', () => {
-    const value = null;
-    const targetUnit = 'in';
-
-    convert(value, targetUnit);
-
-    expect(math.unit).not.toBeCalled();
-    expect(mockMathUnitToNumber).not.toBeCalled();
-  });
-
-  it('returns input value directly if given units are the same', () => {
-    const value = {
-      value: '70',
-      unit: 'in',
-    };
-    const targetUnit = 'in';
-
-    expect(convert(value, targetUnit)).toEqual(70);
-    expect(math.unit).not.toBeCalled();
-    expect(mockMathUnitToNumber).not.toBeCalled();
-  });
-
-  it('cuts off result as precision argument given', () => {
-    mockMathUnitToNumber.mockReturnValue(12.9876);
-    const value = {
-      value: '180',
-      unit: 'cm',
-    };
-    const targetUnit = 'in';
-    const precision = 2;
-
-    expect(convert(value, targetUnit, precision)).toEqual(12.99);
-  });
-});
+import { TextUnitInputComponent } from '../index';
 
 describe('<TextUnitInput />', () => {
   it('can handle decimal point', () => {
