+++ src/web/forms/fields/TextUnitInput/convert.js
@@ -1,0 +1,38 @@
+/**
+ * Copyright 2018 Yuichiro Tsuchiya
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/* @flow */
+
+import math from 'lib/mathjs';
+
+export default (
+  value: ?ValueUnitType,
+  targetUnit: ?string,
+  precision: ?number
+): ?number => {
+  if (!value || !value.value || !value.unit) { return null; }
+  if (!targetUnit) { return null; }
+  if (value.unit === targetUnit) { return parseFloat(value.value); }
+
+  const converted = math.unit(value.value, value.unit).toNumber(targetUnit);
+
+  if (precision != null) {
+    const e = Math.pow(10, precision);
+    return Math.round(converted * e) / e;
+  }
+
+  return converted;
+};
